+++ kscan_test.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"fmt"
+	"kscan/lib/httpfinger"
+	"testing"
+)
+
+func TestName(t *testing.T) {
+	fmt.Println(httpfinger.FaviconHash)
+}
