+++ lib/httpfinger/type-faviconHash.go
@@ -1,0 +1,1994 @@
+package httpfinger
+
+type faviconHash []struct {
+	Hash string `json:"hash"`
+	Cms  string `json:"cms"`
+}
+
+var FaviconHash faviconHash
+
+func (f faviconHash) Match(hash string) string {
+	for _, hSub := range f {
+		if hash == hSub.Hash {
+			return hSub.Cms
+		}
+	}
+	return ""
+}
+
+var faviconHashByte = []byte(`
+[
+    {
+      "hash": "99395752",
+      "cms": "slack-instance"
+    },
+    {
+      "hash": "116323821",
+      "cms": "spring-boot"
+    },
+    {
+      "hash": "81586312",
+      "cms": "Jenkins"
+    },
+    {
+      "hash": "-235701012",
+      "cms": "Cnservers LLC"
+    },
+    {
+      "hash": "743365239",
+      "cms": "Atlassian"
+    },
+    {
+      "hash": "2128230701",
+      "cms": "Chainpoint"
+    },
+    {
+      "hash": "-1277814690",
+      "cms": "LaCie"
+    },
+    {
+      "hash": "246145559",
+      "cms": "Parse"
+    },
+    {
+      "hash": "628535358",
+      "cms": "Atlassian"
+    },
+    {
+      "hash": "855273746",
+      "cms": "JIRA"
+    },
+    {
+      "hash": "1318124267",
+      "cms": "Avigilon"
+    },
+    {
+      "hash": "-305179312",
+      "cms": "Atlassian \u2013 Confluence"
+    },
+    {
+      "hash": "786533217",
+      "cms": "OpenStack"
+    },
+    {
+      "hash": "432733105",
+      "cms": "Pi Star"
+    },
+    {
+      "hash": "705143395",
+      "cms": "Atlassian"
+    },
+    {
+      "hash": "-1255347784",
+      "cms": "Angular IO (AngularJS)"
+    },
+    {
+      "hash": "-1275226814",
+      "cms": "XAMPP"
+    },
+    {
+      "hash": "-2009722838",
+      "cms": "React"
+    },
+    {
+      "hash": "981867722",
+      "cms": "Atlassian \u2013 JIRA"
+    },
+    {
+      "hash": "-923088984",
+      "cms": "OpenStack"
+    },
+    {
+      "hash": "494866796",
+      "cms": "Aplikasi"
+    },
+    {
+      "hash": "1249285083",
+      "cms": "Ubiquiti Aircube"
+    },
+    {
+      "hash": "-1379982221",
+      "cms": "Atlassian \u2013 Bamboo"
+    },
+    {
+      "hash": "420473080",
+      "cms": "Exostar \u2013 Managed Access Gateway"
+    },
+    {
+      "hash": "-1642532491",
+      "cms": "Atlassian \u2013 Confluence"
+    },
+    {
+      "hash": "163842882",
+      "cms": "Cisco Meraki"
+    },
+    {
+      "hash": "-1378182799",
+      "cms": "Archivematica"
+    },
+    {
+      "hash": "-702384832",
+      "cms": "TCN"
+    },
+    {
+      "hash": "-532394952",
+      "cms": "CX"
+    },
+    {
+      "hash": "-183163807",
+      "cms": "Ace"
+    },
+    {
+      "hash": "552727997",
+      "cms": "Atlassian \u2013 JIRA"
+    },
+    {
+      "hash": "1302486561",
+      "cms": "NetData"
+    },
+    {
+      "hash": "-609520537",
+      "cms": "OpenGeo Suite"
+    },
+    {
+      "hash": "-1961046099",
+      "cms": "Dgraph Ratel"
+    },
+    {
+      "hash": "-1581907337",
+      "cms": "Atlassian \u2013 JIRA"
+    },
+    {
+      "hash": "1913538826",
+      "cms": "Material Dashboard"
+    },
+    {
+      "hash": "1319699698",
+      "cms": "Form.io"
+    },
+    {
+      "hash": "-1203021870",
+      "cms": "Kubeflow"
+    },
+    {
+      "hash": "-182423204",
+      "cms": "netdata dashboard"
+    },
+    {
+      "hash": "988422585",
+      "cms": "CapRover"
+    },
+    {
+      "hash": "2113497004",
+      "cms": "WiJungle"
+    },
+    {
+      "hash": "1234311970",
+      "cms": "Onera"
+    },
+    {
+      "hash": "430582574",
+      "cms": "SmartPing"
+    },
+    {
+      "hash": "1232596212",
+      "cms": "OpenStack"
+    },
+    {
+      "hash": "1585145626",
+      "cms": "netdata dashboard"
+    },
+    {
+      "hash": "-219752612",
+      "cms": "FRITZ!Box"
+    },
+    {
+      "hash": "-697231354",
+      "cms": "Ubiquiti \u2013 AirOS"
+    },
+    {
+      "hash": "945408572",
+      "cms": "Fortinet \u2013 Forticlient"
+    },
+    {
+      "hash": "1768726119",
+      "cms": "Outlook Web Application"
+    },
+    {
+      "hash": "2109473187",
+      "cms": "Huawei \u2013 Claro"
+    },
+    {
+      "hash": "552592949",
+      "cms": "ASUS AiCloud"
+    },
+    {
+      "hash": "631108382",
+      "cms": "SonicWALL"
+    },
+    {
+      "hash": "708578229",
+      "cms": "Google"
+    },
+    {
+      "hash": "-134375033",
+      "cms": "Plesk"
+    },
+    {
+      "hash": "2019488876",
+      "cms": "Dahua Storm (IP Camera)"
+    },
+    {
+      "hash": "-1395400951",
+      "cms": "Huawei \u2013 ADSL/Router"
+    },
+    {
+      "hash": "1601194732",
+      "cms": "Sophos Cyberoam (appliance)"
+    },
+    {
+      "hash": "-325082670",
+      "cms": "LANCOM Systems"
+    },
+    {
+      "hash": "-1050786453",
+      "cms": "Plesk"
+    },
+    {
+      "hash": "-1346447358",
+      "cms": "TilginAB (HomeGateway)"
+    },
+    {
+      "hash": "1410610129",
+      "cms": "Supermicro Intelligent Management (IPMI)"
+    },
+    {
+      "hash": "-440644339",
+      "cms": "Zyxel ZyWALL"
+    },
+    {
+      "hash": "363324987",
+      "cms": "Dell SonicWALL"
+    },
+    {
+      "hash": "-1446794564",
+      "cms": "Ubiquiti Login Portals"
+    },
+    {
+      "hash": "1045696447",
+      "cms": "Sophos User Portal/VPN Portal"
+    },
+    {
+      "hash": "-297069493",
+      "cms": "Apache Tomcat"
+    },
+    {
+      "hash": "396533629",
+      "cms": "OpenVPN"
+    },
+    {
+      "hash": "1462981117",
+      "cms": "Cyberoam"
+    },
+    {
+      "hash": "1772087922",
+      "cms": "ASP.net favicon"
+    },
+    {
+      "hash": "1594377337",
+      "cms": "Technicolor"
+    },
+    {
+      "hash": "165976831",
+      "cms": "Vodafone (Technicolor)"
+    },
+    {
+      "hash": "-1677255344",
+      "cms": "UBNT Router UI"
+    },
+    {
+      "hash": "-359621743",
+      "cms": "Intelbras Wireless"
+    },
+    {
+      "hash": "-677167908",
+      "cms": "Kerio Connect (Webmail)"
+    },
+    {
+      "hash": "878647854",
+      "cms": "BIG-IP"
+    },
+    {
+      "hash": "442749392",
+      "cms": "Microsoft OWA"
+    },
+    {
+      "hash": "1405460984",
+      "cms": "pfSense"
+    },
+    {
+      "hash": "-271448102",
+      "cms": "iKuai Networks"
+    },
+    {
+      "hash": "31972968",
+      "cms": "Dlink Webcam"
+    },
+    {
+      "hash": "970132176",
+      "cms": "3CX Phone System"
+    },
+    {
+      "hash": "-1119613926",
+      "cms": "Bluehost"
+    },
+    {
+      "hash": "123821839",
+      "cms": "Sangfor"
+    },
+    {
+      "hash": "459900502",
+      "cms": "ZTE Corporation (Gateway/Appliance)"
+    },
+    {
+      "hash": "-2069844696",
+      "cms": "Ruckus Wireless"
+    },
+    {
+      "hash": "-1607644090",
+      "cms": "Bitnami"
+    },
+    {
+      "hash": "2141724739",
+      "cms": "Juniper Device Manager"
+    },
+    {
+      "hash": "1835479497",
+      "cms": "Technicolor Gateway"
+    },
+    {
+      "hash": "1278323681",
+      "cms": "Gitlab"
+    },
+    {
+      "hash": "-1929912510",
+      "cms": "NETASQ - Secure / Stormshield"
+    },
+    {
+      "hash": "-1255992602",
+      "cms": "VMware Horizon"
+    },
+    {
+      "hash": "1895360511",
+      "cms": "VMware Horizon"
+    },
+    {
+      "hash": "-991123252",
+      "cms": "VMware Horizon"
+    },
+    {
+      "hash": "1642701741",
+      "cms": "Vmware Secure File Transfer"
+    },
+    {
+      "hash": "-266008933",
+      "cms": "SAP Netweaver"
+    },
+    {
+      "hash": "-1967743928",
+      "cms": "SAP ID Service: Log On"
+    },
+    {
+      "hash": "1347937389",
+      "cms": "SAP Conversational AI"
+    },
+    {
+      "hash": "602431586",
+      "cms": "Palo Alto Login Portal"
+    },
+    {
+      "hash": "-318947884",
+      "cms": "Palo Alto Networks"
+    },
+    {
+      "hash": "1356662359",
+      "cms": "Outlook Web Application"
+    },
+    {
+      "hash": "1453890729",
+      "cms": "Webmin"
+    },
+    {
+      "hash": "-1814887000",
+      "cms": "Docker"
+    },
+    {
+      "hash": "1937209448",
+      "cms": "Docker"
+    },
+    {
+      "hash": "-1544605732",
+      "cms": "Amazon"
+    },
+    {
+      "hash": "716989053",
+      "cms": "Amazon"
+    },
+    {
+      "hash": "-1010568750",
+      "cms": "phpMyAdmin"
+    },
+    {
+      "hash": "-1240222446",
+      "cms": "Zhejiang Uniview Technologies Co."
+    },
+    {
+      "hash": "-986678507",
+      "cms": "ISP Manager"
+    },
+    {
+      "hash": "-1616143106",
+      "cms": "AXIS (network cameras)"
+    },
+    {
+      "hash": "-976235259",
+      "cms": "Roundcube Webmail"
+    },
+    {
+      "hash": "768816037",
+      "cms": "UniFi Video Controller (airVision)"
+    },
+    {
+      "hash": "1015545776",
+      "cms": "pfSense"
+    },
+    {
+      "hash": "1838417872",
+      "cms": "Freebox OS"
+    },
+    {
+      "hash": "547282364",
+      "cms": "Keenetic"
+    },
+    {
+      "hash": "-1571472432",
+      "cms": "Sierra Wireless Ace Manager (Airlink)"
+    },
+    {
+      "hash": "149371702",
+      "cms": "Synology DiskStation"
+    },
+    {
+      "hash": "-1169314298",
+      "cms": "INSTAR IP Cameras"
+    },
+    {
+      "hash": "-1038557304",
+      "cms": "Webmin"
+    },
+    {
+      "hash": "1307375944",
+      "cms": "Octoprint (3D printer)"
+    },
+    {
+      "hash": "1280907310",
+      "cms": "Webmin"
+    },
+    {
+      "hash": "1954835352",
+      "cms": "Vesta Hosting Control Panel"
+    },
+    {
+      "hash": "509789953",
+      "cms": "Farming Simulator Dedicated Server"
+    },
+    {
+      "hash": "-1933493443",
+      "cms": "Residential Gateway"
+    },
+    {
+      "hash": "1993518473",
+      "cms": "cPanel Login"
+    },
+    {
+      "hash": "-1477563858",
+      "cms": "Arris"
+    },
+    {
+      "hash": "-895890586",
+      "cms": "PLEX Server"
+    },
+    {
+      "hash": "-1354933624",
+      "cms": "Dlink Webcam"
+    },
+    {
+      "hash": "944969688",
+      "cms": "Deluge"
+    },
+    {
+      "hash": "479413330",
+      "cms": "Webmin"
+    },
+    {
+      "hash": "-435817905",
+      "cms": "Cambium Networks"
+    },
+    {
+      "hash": "-981606721",
+      "cms": "Plesk"
+    },
+    {
+      "hash": "833190513",
+      "cms": "Dahua Storm (IP Camera)"
+    },
+    {
+      "hash": "-652508439",
+      "cms": "Parallels Plesk Panel"
+    },
+    {
+      "hash": "-569941107",
+      "cms": "Fireware Watchguard"
+    },
+    {
+      "hash": "1326164945",
+      "cms": "Shock&Innovation!! netis setup"
+    },
+    {
+      "hash": "-1738184811",
+      "cms": "cacaoweb"
+    },
+    {
+      "hash": "904434662",
+      "cms": "Loxone (Automation)"
+    },
+    {
+      "hash": "905744673",
+      "cms": "HP Printer / Server"
+    },
+    {
+      "hash": "902521196",
+      "cms": "Netflix"
+    },
+    {
+      "hash": "-2063036701",
+      "cms": "Linksys Smart Wi-Fi"
+    },
+    {
+      "hash": "-1205024243",
+      "cms": "lwIP (A Lightweight TCP/IP stack)"
+    },
+    {
+      "hash": "607846949",
+      "cms": "Hitron Technologies"
+    },
+    {
+      "hash": "1281253102",
+      "cms": "Dahua Storm (DVR)"
+    },
+    {
+      "hash": "661332347",
+      "cms": "MOBOTIX Camera"
+    },
+    {
+      "hash": "-520888198",
+      "cms": "Blue Iris (Webcam)"
+    },
+    {
+      "hash": "104189364",
+      "cms": "Vigor Router"
+    },
+    {
+      "hash": "1227052603",
+      "cms": "Alibaba Cloud (Block Page)"
+    },
+    {
+      "hash": "252728887",
+      "cms": "DD WRT (DD-WRT milli_httpd)"
+    },
+    {
+      "hash": "-1922044295",
+      "cms": "Mitel Networks (MiCollab End User Portal)"
+    },
+    {
+      "hash": "1221759509",
+      "cms": "Dlink Webcam"
+    },
+    {
+      "hash": "1037387972",
+      "cms": "Dlink Router"
+    },
+    {
+      "hash": "-655683626",
+      "cms": "PRTG Network Monitor"
+    },
+    {
+      "hash": "1611729805",
+      "cms": "Elastic (Database)"
+    },
+    {
+      "hash": "1144925962",
+      "cms": "Dlink Webcam"
+    },
+    {
+      "hash": "-1666561833",
+      "cms": "Wildfly"
+    },
+    {
+      "hash": "804949239",
+      "cms": "Cisco Meraki Dashboard"
+    },
+    {
+      "hash": "-459291760",
+      "cms": "Workday"
+    },
+    {
+      "hash": "1734609466",
+      "cms": "JustHost"
+    },
+    {
+      "hash": "-1507567067",
+      "cms": "Baidu (IP error page)"
+    },
+    {
+      "hash": "2006716043",
+      "cms": "Intelbras SA"
+    },
+    {
+      "hash": "-1298108480",
+      "cms": "Yii PHP Framework (Default Favicon)"
+    },
+    {
+      "hash": "1782271534",
+      "cms": "truVision NVR (interlogix)"
+    },
+    {
+      "hash": "603314",
+      "cms": "Redmine"
+    },
+    {
+      "hash": "-476231906",
+      "cms": "phpMyAdmin"
+    },
+    {
+      "hash": "-646322113",
+      "cms": "Cisco (eg:Conference Room Login Page)"
+    },
+    {
+      "hash": "-629047854",
+      "cms": "Jetty 404"
+    },
+    {
+      "hash": "-1351901211",
+      "cms": "Luma Surveillance"
+    },
+    {
+      "hash": "-519765377",
+      "cms": "Parallels Plesk Panel"
+    },
+    {
+      "hash": "-2144363468",
+      "cms": "HP Printer / Server"
+    },
+    {
+      "hash": "-127886975",
+      "cms": "Metasploit"
+    },
+    {
+      "hash": "1139788073",
+      "cms": "Metasploit"
+    },
+    {
+      "hash": "-1235192469",
+      "cms": "Metasploit"
+    },
+    {
+      "hash": "1876585825",
+      "cms": "ALIBI NVR"
+    },
+    {
+      "hash": "-1810847295",
+      "cms": "Sangfor"
+    },
+    {
+      "hash": "-291579889",
+      "cms": "Websockets test page (eg: port 5900)"
+    },
+    {
+      "hash": "1629518721",
+      "cms": "macOS Server (Apple)"
+    },
+    {
+      "hash": "-986816620",
+      "cms": "OpenRG"
+    },
+    {
+      "hash": "-299287097",
+      "cms": "Cisco Router"
+    },
+    {
+      "hash": "-1926484046",
+      "cms": "Sangfor"
+    },
+    {
+      "hash": "-873627015",
+      "cms": "HeroSpeed Digital Technology Co. (NVR/IPC/XVR)"
+    },
+    {
+      "hash": "2071993228",
+      "cms": "Nomadix Access Gateway"
+    },
+    {
+      "hash": "516963061",
+      "cms": "Gitlab"
+    },
+    {
+      "hash": "-38580010",
+      "cms": "Magento"
+    },
+    {
+      "hash": "1490343308",
+      "cms": "MK-AUTH"
+    },
+    {
+      "hash": "-632583950",
+      "cms": "Shoutcast Server"
+    },
+    {
+      "hash": "95271369",
+      "cms": "FireEye"
+    },
+    {
+      "hash": "1476335317",
+      "cms": "FireEye"
+    },
+    {
+      "hash": "-842192932",
+      "cms": "FireEye"
+    },
+    {
+      "hash": "105083909",
+      "cms": "FireEye"
+    },
+    {
+      "hash": "240606739",
+      "cms": "FireEye"
+    },
+    {
+      "hash": "2121539357",
+      "cms": "FireEye"
+    },
+    {
+      "hash": "-333791179",
+      "cms": "Adobe Campaign Classic"
+    },
+    {
+      "hash": "-1437701105",
+      "cms": "XAMPP"
+    },
+    {
+      "hash": "-676077969",
+      "cms": "Niagara Web Server"
+    },
+    {
+      "hash": "-2138771289",
+      "cms": "Technicolor"
+    },
+    {
+      "hash": "711742418",
+      "cms": "Hitron Technologies Inc."
+    },
+    {
+      "hash": "728788645",
+      "cms": "IBM Notes"
+    },
+    {
+      "hash": "1436966696",
+      "cms": "Barracuda"
+    },
+    {
+      "hash": "86919334",
+      "cms": "ServiceNow"
+    },
+    {
+      "hash": "1211608009",
+      "cms": "Openfire Admin Console"
+    },
+    {
+      "hash": "2059618623",
+      "cms": "HP iLO"
+    },
+    {
+      "hash": "1975413433",
+      "cms": "Sunny WebBox"
+    },
+    {
+      "hash": "943925975",
+      "cms": "ZyXEL"
+    },
+    {
+      "hash": "281559989",
+      "cms": "Huawei"
+    },
+    {
+      "hash": "-2145085239",
+      "cms": "Tenda Web Master"
+    },
+    {
+      "hash": "-1399433489",
+      "cms": "Prometheus Time Series Collection and Processing Server"
+    },
+    {
+      "hash": "1786752597",
+      "cms": "wdCP cloud host management system"
+    },
+    {
+      "hash": "90680708",
+      "cms": "Domoticz (Home Automation)"
+    },
+    {
+      "hash": "-1441956789",
+      "cms": "Tableau"
+    },
+    {
+      "hash": "-675839242",
+      "cms": "openWRT Luci"
+    },
+    {
+      "hash": "1020814938",
+      "cms": "Ubiquiti \u2013 AirOS"
+    },
+    {
+      "hash": "-766957661",
+      "cms": "MDaemon Webmail"
+    },
+    {
+      "hash": "119741608",
+      "cms": "Teltonika"
+    },
+    {
+      "hash": "1973665246",
+      "cms": "Entrolink"
+    },
+    {
+      "hash": "74935566",
+      "cms": "WindRiver-WebServer"
+    },
+    {
+      "hash": "-1723752240",
+      "cms": "Microhard Systems"
+    },
+    {
+      "hash": "-1807411396",
+      "cms": "Skype"
+    },
+    {
+      "hash": "-1612496354",
+      "cms": "Teltonika"
+    },
+    {
+      "hash": "1877797890",
+      "cms": "Eltex (Router)"
+    },
+    {
+      "hash": "-375623619",
+      "cms": "bintec elmeg"
+    },
+    {
+      "hash": "1483097076",
+      "cms": "SyncThru Web Service (Printers)"
+    },
+    {
+      "hash": "1169183049",
+      "cms": "BoaServer"
+    },
+    {
+      "hash": "1051648103",
+      "cms": "Securepoint"
+    },
+    {
+      "hash": "-438482901",
+      "cms": "Moodle"
+    },
+    {
+      "hash": "-1492966240",
+      "cms": "RADIX"
+    },
+    {
+      "hash": "1466912879",
+      "cms": "CradlePoint Technology (Router)"
+    },
+    {
+      "hash": "-167656799",
+      "cms": "Drupal"
+    },
+    {
+      "hash": "-1593651747",
+      "cms": "Blackboard"
+    },
+    {
+      "hash": "-895963602",
+      "cms": "Jupyter Notebook"
+    },
+    {
+      "hash": "-972810761",
+      "cms": "HostMonster - Web hosting"
+    },
+    {
+      "hash": "1703788174",
+      "cms": "D-Link (router/network)"
+    },
+    {
+      "hash": "225632504",
+      "cms": "Rocket Chat"
+    },
+    {
+      "hash": "-1702393021",
+      "cms": "mofinetwork"
+    },
+    {
+      "hash": "892542951",
+      "cms": "Zabbix"
+    },
+    {
+      "hash": "547474373",
+      "cms": "TOTOLINK (network)"
+    },
+    {
+      "hash": "-374235895",
+      "cms": "Ossia (Provision SR) | Webcam/IP Camera"
+    },
+    {
+      "hash": "1544230796",
+      "cms": "cPanel Login"
+    },
+    {
+      "hash": "517158172",
+      "cms": "D-Link (router/network)"
+    },
+    {
+      "hash": "462223993",
+      "cms": "Jeedom (home automation)"
+    },
+    {
+      "hash": "937999361",
+      "cms": "JBoss Application Server 7"
+    },
+    {
+      "hash": "1991562061",
+      "cms": "Niagara Web Server / Tridium"
+    },
+    {
+      "hash": "812385209",
+      "cms": "Solarwinds Serv-U FTP Server"
+    },
+    {
+      "hash": "1142227528",
+      "cms": "Aruba (Virtual Controller)"
+    },
+    {
+      "hash": "-1153950306",
+      "cms": "Dell"
+    },
+    {
+      "hash": "72005642",
+      "cms": "RemObjects SDK / Remoting SDK for .NET HTTP Server Microsoft"
+    },
+    {
+      "hash": "-484708885",
+      "cms": "Zyxel ZyWALL"
+    },
+    {
+      "hash": "706602230",
+      "cms": "VisualSVN Server"
+    },
+    {
+      "hash": "-656811182",
+      "cms": "Jboss"
+    },
+    {
+      "hash": "-332324409",
+      "cms": "STARFACE VoIP Software"
+    },
+    {
+      "hash": "-594256627",
+      "cms": "Netis (network devices)"
+    },
+    {
+      "hash": "-649378830",
+      "cms": "WHM"
+    },
+    {
+      "hash": "97604680",
+      "cms": "Tandberg"
+    },
+    {
+      "hash": "-1015932800",
+      "cms": "Ghost (CMS)"
+    },
+    {
+      "hash": "-194439630",
+      "cms": "Avtech IP Surveillance (Camera)"
+    },
+    {
+      "hash": "129457226",
+      "cms": "Liferay Portal"
+    },
+    {
+      "hash": "-771764544",
+      "cms": "Parallels Plesk Panel"
+    },
+    {
+      "hash": "-617743584",
+      "cms": "Odoo"
+    },
+    {
+      "hash": "77044418",
+      "cms": "Polycom"
+    },
+    {
+      "hash": "980692677",
+      "cms": "Cake PHP"
+    },
+    {
+      "hash": "476213314",
+      "cms": "Exacq"
+    },
+    {
+      "hash": "794809961",
+      "cms": "CheckPoint"
+    },
+    {
+      "hash": "1157789622",
+      "cms": "Ubiquiti UNMS"
+    },
+    {
+      "hash": "1244636413",
+      "cms": "cPanel Login"
+    },
+    {
+      "hash": "1985721423",
+      "cms": "WorldClient for Mdaemon"
+    },
+    {
+      "hash": "-1124868062",
+      "cms": "Netport Software (DSL)"
+    },
+    {
+      "hash": "-335242539",
+      "cms": "f5 Big IP"
+    },
+    {
+      "hash": "2146763496",
+      "cms": "Mailcow"
+    },
+    {
+      "hash": "-1041180225",
+      "cms": "QNAP NAS Virtualization Station"
+    },
+    {
+      "hash": "-1319025408",
+      "cms": "Netgear"
+    },
+    {
+      "hash": "917966895",
+      "cms": "Gogs"
+    },
+    {
+      "hash": "512590457",
+      "cms": "Trendnet IP camera"
+    },
+    {
+      "hash": "1678170702",
+      "cms": "Asustor"
+    },
+    {
+      "hash": "-1466785234",
+      "cms": "Dahua"
+    },
+    {
+      "hash": "-505448917",
+      "cms": "Discuz!"
+    },
+    {
+      "hash": "255892555",
+      "cms": "wdCP cloud host management system"
+    },
+    {
+      "hash": "1627330242",
+      "cms": "Joomla"
+    },
+    {
+      "hash": "-1935525788",
+      "cms": "SmarterMail"
+    },
+    {
+      "hash": "-12700016",
+      "cms": "Seafile"
+    },
+    {
+      "hash": "1770799630",
+      "cms": "bintec elmeg"
+    },
+    {
+      "hash": "-137295400",
+      "cms": "NETGEAR ReadyNAS"
+    },
+    {
+      "hash": "-195508437",
+      "cms": "iPECS"
+    },
+    {
+      "hash": "-2116540786",
+      "cms": "bet365"
+    },
+    {
+      "hash": "-38705358",
+      "cms": "Reolink"
+    },
+    {
+      "hash": "-450254253",
+      "cms": "idera"
+    },
+    {
+      "hash": "-1630354993",
+      "cms": "Proofpoint"
+    },
+    {
+      "hash": "-1678298769",
+      "cms": "Kerio Connect WebMail"
+    },
+    {
+      "hash": "-35107086",
+      "cms": "WorldClient for Mdaemon"
+    },
+    {
+      "hash": "2055322029",
+      "cms": "Realtek"
+    },
+    {
+      "hash": "-692947551",
+      "cms": "Ruijie Networks (Login)"
+    },
+    {
+      "hash": "-1710631084",
+      "cms": "Askey Cable Modem"
+    },
+    {
+      "hash": "89321398",
+      "cms": "Askey Cable Modem"
+    },
+    {
+      "hash": "90066852",
+      "cms": "JAWS Web Server (IP Camera)"
+    },
+    {
+      "hash": "768231242",
+      "cms": "JAWS Web Server (IP Camera)"
+    },
+    {
+      "hash": "-421986013",
+      "cms": "Homegrown Website Hosting"
+    },
+    {
+      "hash": "156312019",
+      "cms": "Technicolor / Thomson Speedtouch (Network / ADSL)"
+    },
+    {
+      "hash": "-560297467",
+      "cms": "DVR (Korean)"
+    },
+    {
+      "hash": "-1950415971",
+      "cms": "Joomla"
+    },
+    {
+      "hash": "1842351293",
+      "cms": "TP-LINK (Network Device)"
+    },
+    {
+      "hash": "1433417005",
+      "cms": "Salesforce"
+    },
+    {
+      "hash": "-632070065",
+      "cms": "Apache Haus"
+    },
+    {
+      "hash": "1103599349",
+      "cms": "Untangle"
+    },
+    {
+      "hash": "224536051",
+      "cms": "Shenzhen coship electronics co."
+    },
+    {
+      "hash": "1038500535",
+      "cms": "D-Link (router/network)"
+    },
+    {
+      "hash": "-355305208",
+      "cms": "D-Link (camera)"
+    },
+    {
+      "hash": "-267431135",
+      "cms": "Kibana"
+    },
+    {
+      "hash": "-759754862",
+      "cms": "Kibana"
+    },
+    {
+      "hash": "-1200737715",
+      "cms": "Kibana"
+    },
+    {
+      "hash": "75230260",
+      "cms": "Kibana"
+    },
+    {
+      "hash": "1668183286",
+      "cms": "Kibana"
+    },
+    {
+      "hash": "283740897",
+      "cms": "Intelbras SA"
+    },
+    {
+      "hash": "1424295654",
+      "cms": "Icecast Streaming Media Server"
+    },
+    {
+      "hash": "1922032523",
+      "cms": "NEC WebPro"
+    },
+    {
+      "hash": "-1654229048",
+      "cms": "Vivotek (Camera)"
+    },
+    {
+      "hash": "-1414475558",
+      "cms": "Microsoft IIS"
+    },
+    {
+      "hash": "-1697334194",
+      "cms": "Univention Portal"
+    },
+    {
+      "hash": "-1424036600",
+      "cms": "Portainer (Docker Management)"
+    },
+    {
+      "hash": "-831826827",
+      "cms": "NOS Router"
+    },
+    {
+      "hash": "-759108386",
+      "cms": "Tongda"
+    },
+    {
+      "hash": "-1022206565",
+      "cms": "CrushFTP"
+    },
+    {
+      "hash": "-1225484776",
+      "cms": "Endian Firewall"
+    },
+    {
+      "hash": "-631002664",
+      "cms": "Kerio Control Firewall"
+    },
+    {
+      "hash": "2072198544",
+      "cms": "Ferozo Panel"
+    },
+    {
+      "hash": "-466504476",
+      "cms": "Kerio Control Firewall"
+    },
+    {
+      "hash": "1251810433",
+      "cms": "Cafe24 (Korea)"
+    },
+    {
+      "hash": "1273982002",
+      "cms": "Mautic (Open Source Marketing Automation)"
+    },
+    {
+      "hash": "-978656757",
+      "cms": "NETIASPOT (Network)"
+    },
+    {
+      "hash": "916642917",
+      "cms": "Multilaser"
+    },
+    {
+      "hash": "575613323",
+      "cms": "Canvas LMS (Learning Management)"
+    },
+    {
+      "hash": "1726027799",
+      "cms": "IBM Server"
+    },
+    {
+      "hash": "-587741716",
+      "cms": "ADB Broadband S.p.A. (Network)"
+    },
+    {
+      "hash": "-360566773",
+      "cms": "ARRIS (Network)"
+    },
+    {
+      "hash": "-884776764",
+      "cms": "Huawei (Network)"
+    },
+    {
+      "hash": "929825723",
+      "cms": "WAMPSERVER"
+    },
+    {
+      "hash": "240136437",
+      "cms": "Seagate Technology (NAS)"
+    },
+    {
+      "hash": "1911253822",
+      "cms": "UPC Ceska Republica (Network)"
+    },
+    {
+      "hash": "-393788031",
+      "cms": "Flussonic (Video Streaming)"
+    },
+    {
+      "hash": "366524387",
+      "cms": "Joomla"
+    },
+    {
+      "hash": "443944613",
+      "cms": "WAMPSERVER"
+    },
+    {
+      "hash": "1953726032",
+      "cms": "Metabase"
+    },
+    {
+      "hash": "-2031183903",
+      "cms": "D-Link (Network)"
+    },
+    {
+      "hash": "545827989",
+      "cms": "MobileIron"
+    },
+    {
+      "hash": "967636089",
+      "cms": "MobileIron"
+    },
+    {
+      "hash": "362091310",
+      "cms": "MobileIron"
+    },
+    {
+      "hash": "2086228042",
+      "cms": "MobileIron"
+    },
+    {
+      "hash": "-1588746893",
+      "cms": "CommuniGate"
+    },
+    {
+      "hash": "1427976651",
+      "cms": "ZTE (Network)"
+    },
+    {
+      "hash": "1648531157",
+      "cms": "InfiNet Wireless | WANFleX (Network)"
+    },
+    {
+      "hash": "938616453",
+      "cms": "Mersive Solstice"
+    },
+    {
+      "hash": "1632780968",
+      "cms": "Universit\u00e9 Toulouse 1 Capitole"
+    },
+    {
+      "hash": "2068154487",
+      "cms": "Digium (Switchvox)"
+    },
+    {
+      "hash": "-1788112745",
+      "cms": "PowerMTA monitoring"
+    },
+    {
+      "hash": "-644617577",
+      "cms": "SmartLAN/G"
+    },
+    {
+      "hash": "-1822098181",
+      "cms": "Checkpoint (Gaia)"
+    },
+    {
+      "hash": "-1131689409",
+      "cms": "\u0423\u0422\u041c (Federal Service for Alcohol Market Regulation | Russia)"
+    },
+    {
+      "hash": "2127152956",
+      "cms": "MailWizz"
+    },
+    {
+      "hash": "1064742722",
+      "cms": "RabbitMQ"
+    },
+    {
+      "hash": "-693082538",
+      "cms": "openmediavault (NAS)"
+    },
+    {
+      "hash": "1941381095",
+      "cms": "openWRT Luci"
+    },
+    {
+      "hash": "903086190",
+      "cms": "Honeywell"
+    },
+    {
+      "hash": "829321644",
+      "cms": "BOMGAR Support Portal"
+    },
+    {
+      "hash": "-1442789563",
+      "cms": "Nuxt JS"
+    },
+    {
+      "hash": "-2140379067",
+      "cms": "RoundCube Webmail"
+    },
+    {
+      "hash": "-1897829998",
+      "cms": "D-Link (camera)"
+    },
+    {
+      "hash": "1047213685",
+      "cms": "Netgear (Network)"
+    },
+    {
+      "hash": "1485257654",
+      "cms": "SonarQube"
+    },
+    {
+      "hash": "-299324825",
+      "cms": "Lupus Electronics XT"
+    },
+    {
+      "hash": "-1162730477",
+      "cms": "Vanderbilt SPC"
+    },
+    {
+      "hash": "-1268095485",
+      "cms": "VZPP Plesk"
+    },
+    {
+      "hash": "1118684072",
+      "cms": "Baidu"
+    },
+    {
+      "hash": "-1616115760",
+      "cms": "ownCloud"
+    },
+    {
+      "hash": "-2054889066",
+      "cms": "Sentora"
+    },
+    {
+      "hash": "1333537166",
+      "cms": "Alfresco"
+    },
+    {
+      "hash": "-373674173",
+      "cms": "Digital Keystone (DK)"
+    },
+    {
+      "hash": "-106646451",
+      "cms": "WISPR (Airlan)"
+    },
+    {
+      "hash": "1235070469",
+      "cms": "Synology VPN Plus"
+    },
+    {
+      "hash": "2063428236",
+      "cms": "Sentry"
+    },
+    {
+      "hash": "15831193",
+      "cms": "WatchGuard"
+    },
+    {
+      "hash": "-956471263",
+      "cms": "Web Client Pro"
+    },
+    {
+      "hash": "-1452159623",
+      "cms": "Tecvoz"
+    },
+    {
+      "hash": "99432374",
+      "cms": "MDaemon Remote Administration"
+    },
+    {
+      "hash": "727253975",
+      "cms": "Paradox IP Module"
+    },
+    {
+      "hash": "-630493013",
+      "cms": "DokuWiki"
+    },
+    {
+      "hash": "552597979",
+      "cms": "Sails"
+    },
+    {
+      "hash": "774252049",
+      "cms": "FastPanel Hosting"
+    },
+    {
+      "hash": "-329747115",
+      "cms": "C-Lodop"
+    },
+    {
+      "hash": "1262005940",
+      "cms": "Jamf Pro Login"
+    },
+    {
+      "hash": "979634648",
+      "cms": "StruxureWare (Schneider Electric)"
+    },
+    {
+      "hash": "475379699",
+      "cms": "Axcient Replibit Management Server"
+    },
+    {
+      "hash": "-878891718",
+      "cms": "Twonky Server (Media Streaming)"
+    },
+    {
+      "hash": "-2125083197",
+      "cms": "Windows Azure"
+    },
+    {
+      "hash": "-1151675028",
+      "cms": "ISP Manager (Web Hosting Panel)"
+    },
+    {
+      "hash": "1248917303",
+      "cms": "JupyterHub"
+    },
+    {
+      "hash": "-1908556829",
+      "cms": "CenturyLink Modem GUI Login (eg: Technicolor)"
+    },
+    {
+      "hash": "1059329877",
+      "cms": "Tecvoz"
+    },
+    {
+      "hash": "-1148190371",
+      "cms": "OPNsense"
+    },
+    {
+      "hash": "1467395679",
+      "cms": "Ligowave (network)"
+    },
+    {
+      "hash": "-1528414776",
+      "cms": "Rumpus"
+    },
+    {
+      "hash": "-2117390767",
+      "cms": "Spiceworks (panel)"
+    },
+    {
+      "hash": "-1944119648",
+      "cms": "TeamCity"
+    },
+    {
+      "hash": "-1748763891",
+      "cms": "INSTAR Full-HD IP-Camera"
+    },
+    {
+      "hash": "251106693",
+      "cms": "GPON Home Gateway"
+    },
+    {
+      "hash": "-1779611449",
+      "cms": "Alienvault"
+    },
+    {
+      "hash": "-1745552996",
+      "cms": "Arbor Networks"
+    },
+    {
+      "hash": "-1275148624",
+      "cms": "Accrisoft"
+    },
+    {
+      "hash": "-178685903",
+      "cms": "Yasni"
+    },
+    {
+      "hash": "-43161126",
+      "cms": "Slack"
+    },
+    {
+      "hash": "671221099",
+      "cms": "innovaphone"
+    },
+    {
+      "hash": "-10974981",
+      "cms": "Shinobi (CCTV)"
+    },
+    {
+      "hash": "1274078387",
+      "cms": "TP-LINK (Network Device)"
+    },
+    {
+      "hash": "-336242473",
+      "cms": "Siemens OZW772"
+    },
+    {
+      "hash": "882208493",
+      "cms": "Lantronix (Spider)"
+    },
+    {
+      "hash": "-687783882",
+      "cms": "ClaimTime (Ramsell Public Health & Safety)"
+    },
+    {
+      "hash": "-590892202",
+      "cms": "Surfilter SSL VPN Portal"
+    },
+    {
+      "hash": "-50306417",
+      "cms": "Kyocera (Printer)"
+    },
+    {
+      "hash": "784872924",
+      "cms": "Lucee!"
+    },
+    {
+      "hash": "1135165421",
+      "cms": "Ricoh"
+    },
+    {
+      "hash": "926501571",
+      "cms": "Handle Proxy"
+    },
+    {
+      "hash": "579239725",
+      "cms": "Metasploit"
+    },
+    {
+      "hash": "-689902428",
+      "cms": "iomega NAS"
+    },
+    {
+      "hash": "-600508822",
+      "cms": "iomega NAS"
+    },
+    {
+      "hash": "656868270",
+      "cms": "iomega NAS"
+    },
+    {
+      "hash": "-2056503929",
+      "cms": "iomega NAS"
+    },
+    {
+      "hash": "-1656695885",
+      "cms": "iomega NAS"
+    },
+    {
+      "hash": "331870709",
+      "cms": "iomega NAS"
+    },
+    {
+      "hash": "1241049726",
+      "cms": "iomega NAS"
+    },
+    {
+      "hash": "998138196",
+      "cms": "iomega NAS"
+    },
+    {
+      "hash": "322531336",
+      "cms": "iomega NAS"
+    },
+    {
+      "hash": "-401934945",
+      "cms": "iomega NAS"
+    },
+    {
+      "hash": "-613216179",
+      "cms": "iomega NAS"
+    },
+    {
+      "hash": "-276759139",
+      "cms": "Chef Automate"
+    },
+    {
+      "hash": "1862132268",
+      "cms": "Gargoyle Router Management Utility"
+    },
+    {
+      "hash": "-1738727418",
+      "cms": "KeepItSafe Management Console"
+    },
+    {
+      "hash": "-368490461",
+      "cms": "Entronix Energy Management Platform"
+    },
+    {
+      "hash": "1836828108",
+      "cms": "OpenProject"
+    },
+    {
+      "hash": "-1775553655",
+      "cms": "Unified Management Console (Polycom)"
+    },
+    {
+      "hash": "381100274",
+      "cms": "Moxapass ioLogik Remote Ethernet I/O Server "
+    },
+    {
+      "hash": "2124459909",
+      "cms": "HFS (HTTP File Server)"
+    },
+    {
+      "hash": "731374291",
+      "cms": "HFS (HTTP File Server)"
+    },
+    {
+      "hash": "-335153896",
+      "cms": "Traccar GPS tracking"
+    },
+    {
+      "hash": "896412703",
+      "cms": "IW"
+    },
+    {
+      "hash": "191654058",
+      "cms": "Wordpress Under Construction Icon"
+    },
+    {
+      "hash": "-342262483",
+      "cms": "Combivox"
+    },
+    {
+      "hash": "5542029",
+      "cms": "NetComWireless (Network)"
+    },
+    {
+      "hash": "1552860581",
+      "cms": "Elastic (Database)"
+    },
+    {
+      "hash": "1174841451",
+      "cms": "Drupal"
+    },
+    {
+      "hash": "-1093172228",
+      "cms": "truVision (NVR)"
+    },
+    {
+      "hash": "-1688698891",
+      "cms": "SpamExperts"
+    },
+    {
+      "hash": "-1546574541",
+      "cms": "Sonatype Nexus Repository Manager"
+    },
+    {
+      "hash": "-256828986",
+      "cms": "iDirect Canada (Network Management)"
+    },
+    {
+      "hash": "1966198264",
+      "cms": "OpenERP (now known as Odoo)"
+    },
+    {
+      "hash": "2099342476",
+      "cms": "PKP (OpenJournalSystems) Public Knowledge Project"
+    },
+    {
+      "hash": "541087742",
+      "cms": "LiquidFiles"
+    },
+    {
+      "hash": "-882760066",
+      "cms": "ZyXEL (Network)"
+    },
+    {
+      "hash": "16202868",
+      "cms": "Universal Devices (UD)"
+    },
+    {
+      "hash": "987967490",
+      "cms": "Huawei (Network)"
+    },
+    {
+      "hash": "1969970750",
+      "cms": "Gitea"
+    },
+    {
+      "hash": "-1734573358",
+      "cms": "TC-Group"
+    },
+    {
+      "hash": "-1589842876",
+      "cms": "Deluge Web UI"
+    },
+    {
+      "hash": "1822002133",
+      "cms": "\u767b\u5f55 \u2013 AMH"
+    },
+    {
+      "hash": "-2006308185",
+      "cms": "OTRS (Open Ticket Request System)"
+    },
+    {
+      "hash": "-1702769256",
+      "cms": "Bosch Security Systems (Camera)"
+    },
+    {
+      "hash": "321591353",
+      "cms": "Node-RED"
+    },
+    {
+      "hash": "-923693877",
+      "cms": "motionEye (camera)"
+    },
+    {
+      "hash": "-1547576879",
+      "cms": "Saia Burgess Controls \u2013 PCD"
+    },
+    {
+      "hash": "1479202414",
+      "cms": "Arcadyan o2 box (Network)"
+    },
+    {
+      "hash": "1081719753",
+      "cms": "D-Link (Network)"
+    },
+    {
+      "hash": "-166151761",
+      "cms": "Abilis (Network/Automation)"
+    },
+    {
+      "hash": "-1231681737",
+      "cms": "Ghost (CMS)"
+    },
+    {
+      "hash": "321909464",
+      "cms": "Airwatch"
+    },
+    {
+      "hash": "-1153873472",
+      "cms": "Airwatch"
+    },
+    {
+      "hash": "1095915848",
+      "cms": "Airwatch"
+    },
+    {
+      "hash": "788771792",
+      "cms": "Airwatch"
+    },
+    {
+      "hash": "-1863663974",
+      "cms": "Airwatch"
+    },
+    {
+      "hash": "-1267819858",
+      "cms": "KeyHelp (Keyweb AG)"
+    },
+    {
+      "hash": "726817668",
+      "cms": "KeyHelp (Keyweb AG)"
+    },
+    {
+      "hash": "-1474875778",
+      "cms": "GLPI"
+    },
+    {
+      "hash": "5471989",
+      "cms": "Netcom Technology"
+    },
+    {
+      "hash": "-1457536113",
+      "cms": "CradlePoint"
+    },
+    {
+      "hash": "-736276076",
+      "cms": "MyASP"
+    },
+    {
+      "hash": "-1343070146",
+      "cms": "Intelbras SA"
+    },
+    {
+      "hash": "538585915",
+      "cms": "Lenel"
+    },
+    {
+      "hash": "-625364318",
+      "cms": "OkoFEN Pellematic"
+    },
+    {
+      "hash": "1117165781",
+      "cms": "SimpleHelp (Remote Support)"
+    },
+    {
+      "hash": "-1067420240",
+      "cms": "GraphQL"
+    },
+    {
+      "hash": "-1465479343",
+      "cms": "DNN (CMS)"
+    },
+    {
+      "hash": "1232159009",
+      "cms": "Apple"
+    },
+    {
+      "hash": "1382324298",
+      "cms": "Apple"
+    },
+    {
+      "hash": "-1498185948",
+      "cms": "Apple"
+    },
+    {
+      "hash": "483383992",
+      "cms": "ISPConfig"
+    },
+    {
+      "hash": "-1249852061",
+      "cms": "Microsoft Outlook"
+    },
+    {
+      "hash": "999357577",
+      "cms": "Hikvision IP Camera"
+    },
+    {
+      "hash": "492290497",
+      "cms": "IP Camera"
+    },
+    {
+      "hash": "-194791768",
+      "cms": "AfterLogicWebMail\u7cfb\u7edf"
+    },
+    {
+      "hash": "492941040",
+      "cms": "B2Bbuilder"
+    }
+]
+`)
