+++ lib/iconhash/iconhash.go
