+++ lib/net/stcp/stcp.go
