--- src/app/config/type.go
@@ -1,11 +1,0 @@
-package config
-
-type (
-	//扫描配置
-	config struct {
-	}
-)
-
-//Par
-
-var defaultConfig = config{}
