--- src/lib/queue/queue.go
@@ -1,67 +1,0 @@
-package queue
-
-type (
-	//Queue 队列
-	Queue struct {
-		top    *node
-		rear   *node
-		length int
-	}
-	//双向链表节点
-	node struct {
-		pre   *node
-		next  *node
-		value interface{}
-	}
-)
-
-// Create a new queue
-func New() *Queue {
-	return &Queue{nil, nil, 0}
-}
-
-//获取队列长度
-func (this *Queue) Len() int {
-	return this.length
-}
-
-//返回队列顶端元素
-func (this *Queue) Peek() interface{} {
-	if this.top == nil {
-		return nil
-	}
-	return this.top.value
-}
-
-//入队操作
-func (this *Queue) Push(v interface{}) {
-	n := &node{nil, nil, v}
-	if this.length == 0 {
-		this.top = n
-		this.rear = this.top
-	} else {
-		n.pre = this.rear
-		this.rear.next = n
-		this.rear = n
-	}
-	this.length++
-}
-
-//出队操作
-func (this *Queue) Pop() interface{} {
-	if this.length == 0 {
-		return nil
-	}
-	n := this.top
-	if this.top != nil {
-		if this.top.next == nil {
-			this.top = nil
-		} else {
-			this.top = this.top.next
-			//this.top.pre.next = nil
-			this.top.pre = nil
-		}
-	}
-	this.length--
-	return n.value
-}
