--- src/lib/scan/scan.go
@@ -1,1 +1,0 @@
-package scan
